Restricted a certain Guild from !ripchat.
Upon request of the mods, the Twokinds Community Discord will be denied
access to !ripchat.

+++ index.js
@@ -27,9 +27,15 @@ client.on('message', message => {
       console.log(message.author.tag + ' used ' + message.content);
   }
   if (message.content === prefix + 'ripchatimg') {
+    if (message.guild === '165725395348357120') {
+      return
+    }
+    else {
+
     message.delete()
     message.channel.send('https://prnt.sc/f8lyf9');
     console.log(message.author.tag + ' used ' + message.content);
+    }
   }
   if (message.content === prefix + 'coffee') {
     message.delete()
